% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{adt_cci_plot}
\alias{adt_cci_plot}
\title{Visualization of protein-protein communication}
\usage{
adt_cci_plot(df, color = "#337EDE", mode = "1")
}
\arguments{
\item{df}{Dataframe of protein-protein conmmunication inference}

\item{mode}{}
}
\description{
Visualization of protein-protein communication
}
