% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatrio_resonance.R
\name{weighted_cor}
\alias{weighted_cor}
\title{Spatially weighted correlation}
\usage{
weighted_cor(mat, weighted_mat, method = "pearson", na_zero = T)
}
\arguments{
\item{mat}{Expression matrix, n X p}

\item{weighted_mat}{Weight matrix, n X n}

\item{method}{Correlation method, pearson or spearman}

\item{na_zero}{Na to zero}
}
\value{
Weighted correlation matrix, p X p
}
\description{
Spatially weighted correlation
}
